+++ app/repositories.py
@@ -1,0 +1,2 @@
+class MessageRepository:
+    ...
