+++ app/utils.py
@@ -1,5 +1,8 @@
 import datetime
+import re
 from zoneinfo import ZoneInfo
+
+import googletrans
 
 
 def tz_now(tz: str = "Asia/Seoul") -> datetime.datetime:
@@ -10,3 +13,27 @@ def tz_now(tz: str = "Asia/Seoul") -> datetime.datetime:
 def tz_now_to_str(tz: str = "Asia/Seoul") -> str:
     """현재시간을 문자열로 반환합니다."""
     return datetime.datetime.strftime(tz_now(tz), "%Y-%m-%d %H:%M:%S")
+
+
+def is_english(text):
+    if re.match("^[a-zA-Z]+$", text):
+        return True
+    elif re.match("^[가-힣]+$", text):
+        return False
+    else:
+        return None
+
+
+def translate_keywords(keywords: list[str]) -> list[str]:
+    translator = googletrans.Translator()
+    results = []
+    for keyword in keywords:
+        value = is_english(keyword)
+        if value is True:
+            # 영어 -> 한글 번역, 한글이 없는 단어는 그대로 영어가 나올 수 있음.
+            results.append(translator.translate(keyword, dest="ko").text.lower())
+        elif value is False:
+            results.append(translator.translate(keyword, dest="en").text.lower())
+        else:
+            continue
+    return results
